+++ Raytracer/Scene.h
@@ -14,7 +14,7 @@ public:
 	Scene(vec3 bgColor = vec3((float) 224 / 255, (float) 230 / 255, (float) 237 / 255))
 		:bgColor(bgColor)
 	{}
-	bool Intersect(Ray& ray, float& t, IntersectInfo& info);
+	bool Intersect(Ray& ray, double& t, IntersectInfo& info);
 
 	vector<Sphere> objs;
 	vec3 bgColor;
